small refactorings.
