+++ rhino-archetype/src/main/resources/archetype-resources/src/main/java/benchmark/RhinoEntity.java
@@ -16,9 +16,8 @@
 
 package ${groupId}.benchmark;
 
-import io.ryos.rhino.sdk.users.User;
-
-import io.ryos.rhino.sdk.Recorder;
+import io.ryos.rhino.sdk.users.data.User;
+import io.ryos.rhino.sdk.reporting.Recorder;
 import io.ryos.rhino.sdk.annotations.CleanUp;
 import io.ryos.rhino.sdk.annotations.Feeder;
 import io.ryos.rhino.sdk.annotations.Prepare;
@@ -26,13 +25,11 @@ import io.ryos.rhino.sdk.annotations.Scenario;
 import io.ryos.rhino.sdk.annotations.Simulation;
 import io.ryos.rhino.sdk.annotations.UserFeeder;
 import io.ryos.rhino.sdk.feeders.UUIDFeeder;
-import io.ryos.rhino.sdk.reporting.GatlingLogFormatter;
 
 /**
  * An example for annotated entity of benchmark job.
  */
 @Simulation(name = "helloWorld")
-@Logging(file = "/var/log/simulation.log", formatter = GatlingLogFormatter.class)
 public class RhinoEntity {
 
     @UserFeeder(max = 10)
